Add a `getSelection` method to expose the "selection" property.
The page model JSON exposes the properties from the JCR tree, not the ones exported by the Sling model. The component JSON model (i.e. `producteaser.model.json`) exports the correct property name, as defined in the Sling model (i.e. `sku`). This commit is a workaround that issue until we clarify it.